ai_service/requirements.txt
astapi uvicorn requests pydantic